third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/directive-parsing-03-expected.txt

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'aaa'.

CONSOLE ERROR: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because it violates the following Content Security Policy directive: "script-src 'none'".

This script should not execute even though there are parse errors in the policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
